Basic transplant of keyboard update code.
